Detect and handle duplicated articles from Ritzau [BOR-36]
